refactor(core): move callAndReportToErrorHandler to bootstrap file
This function is only used in the bootstrap file and does not need to be in application_ref
